resume/api/requirements.txt
asgiref==3.5.2
bcrypt==4.0.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.2
django-3-jet==1.0.8
django-colorfield==0.7.2
django-cors-headers==3.13.0
django-environ==0.9.0
django-filter==22.1
django-resized==1.0.2
django-tinymce==3.5.0
django-tinymce4-lite==1.8.0
djangorestframework==3.14.0
djangorestframework-jsonapi==6.0.0
djangorestframework-simplejwt==5.2.1
drf-yasg==1.21.4
environ==1.0
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsmin==3.0.1
MarkupSafe==2.1.1
packaging==21.3
Pillow==9.2.0
psycopg2==2.9.4
pycparser==2.21
pyenchant==3.2.2
PyJWT==2.5.0
pyparsing==3.0.9
pytz==2022.4
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.12
